refactor: move block context menu items into a registry
The 23 items and copiedStyle move to components/BlockContextMenuOptions.
BlockContextMenu keeps showContextMenu, the context object and the
NewBlockTemplate dialog.

Add an all accessor to createRegistry. visible calls condition with no
argument, but a context menu item decides per block, so the menu reads all
and each item applies its own condition to the context.

Split RegistryItem into Identified, which is name and rank, and RegistryItem,
which adds the no-argument condition. The registry now accepts both shapes.

Part of extensibility phase 1, step 6.
